2026/06/19/15:55

+++ src/main/java/com/example/macaron/Storeview.java
@@ -2,7 +2,7 @@ package com.example.macaron;
 
 //パスワード、クーポン、自由記述が編集可能
 public class Storeview {
-    private final Long id;
+    private final Integer id;
     private final String name;
     private final String email;
     private String password;
@@ -14,7 +14,7 @@ public class Storeview {
     private String coupon;
     private String free_desc;
 
-    public Storeview(Long id, String name, String email, String password, String address, Integer phone_number, String payment,
+    public Storeview(Integer id, String name, String email, String password, String address, Integer phone_number, String payment,
             Integer price_range, String genre, String coupon, String free_desc) {
         this.id = id;
         this.name = name;
@@ -29,7 +29,7 @@ public class Storeview {
         this.free_desc = free_desc;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
